pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dash==3.0.3",
    "flask==3.0.3",
    "flask-admin>=1.6.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "pip>=25.0.1",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "wtforms>=3.2.1",
]
